Restructure file to have namespaces in api
Modularized backend to mines, parties, and permits and created namespaces
for each module. This will help with organization and as features
increase there might be a need to break things into its own app or
microservice.
