+++ python-backend/tests/party/resources/test_mgr_appointment_resource.py
@@ -1,13 +1,13 @@
 from datetime import datetime, timedelta
 import json
 
-from app.api.party.models.party import MgrAppointment
+from app.api.parties.party.models.party import MgrAppointment
 from tests.constants import TEST_MINE_GUID, TEST_MANAGER_GUID, TEST_PARTY_PER_GUID_2
 
 
 # GET
 def test_get_manager_not_found(test_client, auth_headers):
-    get_resp = test_client.get('/managers/' + TEST_MINE_GUID, headers=auth_headers['full_auth_header'])
+    get_resp = test_client.get('parties/managers/' + TEST_MINE_GUID, headers=auth_headers['full_auth_header'])
     get_data = json.loads(get_resp.data.decode())
     assert get_data == {
         'error': {
@@ -19,7 +19,7 @@ def test_get_manager_not_found(test_client, auth_headers):
 
 
 def test_get_manager(test_client, auth_headers):
-    get_resp = test_client.get('/managers/' + TEST_MANAGER_GUID, headers=auth_headers['full_auth_header'])
+    get_resp = test_client.get('parties/managers/' + TEST_MANAGER_GUID, headers=auth_headers['full_auth_header'])
     get_data = json.loads(get_resp.data.decode())
     assert get_data['mgr_appointment_guid'] == TEST_MANAGER_GUID
     assert get_resp.status_code == 200
@@ -27,7 +27,7 @@ def test_get_manager(test_client, auth_headers):
 
 # POST
 def test_post_manager_invalid_url(test_client, auth_headers):
-    post_resp = test_client.post('/managers/some_id', data={}, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers/some_id', data={}, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -40,7 +40,7 @@ def test_post_manager_invalid_url(test_client, auth_headers):
 
 def test_post_no_party_guid(test_client, auth_headers):
     test_manager_data = {"mine_guid": TEST_MINE_GUID, "effective_date": datetime.today().strftime("%Y-%m-%d"), "expiry_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -53,7 +53,7 @@ def test_post_no_party_guid(test_client, auth_headers):
 
 def test_post_no_mine_guid(test_client, auth_headers):
     test_manager_data = {"party_guid": TEST_PARTY_PER_GUID_2, "effective_date": datetime.today().strftime("%Y-%m-%d"), "expiry_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -66,7 +66,7 @@ def test_post_no_mine_guid(test_client, auth_headers):
 
 def test_post_no_effective_date(test_client, auth_headers):
     test_manager_data = {"mine_guid": TEST_MINE_GUID, "party_guid": TEST_PARTY_PER_GUID_2, "expiry_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -79,7 +79,7 @@ def test_post_no_effective_date(test_client, auth_headers):
 
 def test_post_person_does_not_exist(test_client, auth_headers):
     test_manager_data = {"mine_guid": TEST_MINE_GUID, "party_guid": TEST_MINE_GUID, "effective_date": datetime.today().strftime("%Y-%m-%d"), "expiry_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -92,7 +92,7 @@ def test_post_person_does_not_exist(test_client, auth_headers):
 
 def test_post_mine_does_not_exist(test_client, auth_headers):
     test_manager_data = {"mine_guid": TEST_PARTY_PER_GUID_2, "party_guid": TEST_PARTY_PER_GUID_2, "effective_date": datetime.today().strftime("%Y-%m-%d"), "expiry_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     assert post_data == {
         'error': {
@@ -105,7 +105,7 @@ def test_post_mine_does_not_exist(test_client, auth_headers):
 
 def test_post_manager_success(test_client, auth_headers):
     test_manager_data = {"party_guid": TEST_PARTY_PER_GUID_2, "mine_guid": TEST_MINE_GUID, "effective_date": datetime.today().strftime("%Y-%m-%d")}
-    post_resp = test_client.post('/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
+    post_resp = test_client.post('parties/managers', data=test_manager_data, headers=auth_headers['full_auth_header'])
     post_data = json.loads(post_resp.data.decode())
     previous_manager = MgrAppointment.find_by_mgr_appointment_guid(TEST_MANAGER_GUID)
     expiry_date = datetime.today() - timedelta(days=1)
